modify directory structure
