+++ testdemo/falsesharing/falsesharing_start.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+for i in {1..100}
+do
+	./falsesharing
+done
