+++ testdemo/multiply/multiply_start.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+for i in {1..200}
+do
+	taskset -c 1 ./multiply
+done
